Massive refactor to create a self-contained, working model1 app.
