+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "libs/Mask_RCNN"]
-	path = libs/Mask_RCNN
+[submodule "src/notebooks/model1_maskrcnn/Mask_RCNN"]
+	path = src/notebooks/model1_maskrcnn/Mask_RCNN
 	url = https://github.com/matterport/Mask_RCNN.git
