+++ setup.cfg
@@ -1,0 +1,14 @@
+[metadata]
+author = Victor W. Yee
+author-email = w.victoryee@gmail.com
+url = https://github.com/victorwyee/lungbox
+keywords = deep learning
+license = Apache
+
+[options]
+zip_safe = False		    # dataset and __init__ use file
+# will call find_packages()
+packages = find:
+
+[wheel]
+universal = 1
