--- src/ingestion/visualize.py
@@ -1,52 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Functions for interacting with AWS"""
-
-import numpy as np
-import pandas as pd
-
-def draw(parsed_df, patient_id):
-    """
-    Method to draw single patient with bounding box(es) if present.
-
-    References:
-    * Exploratory Data Analysis, by Peter Chang, MD.
-        - Available at https://www.kaggle.com/peterchang77/exploratory-data-analysis/notebook
-        - Retrieved 2018-09-19
-    """
-    d = get_s3_dcm(patient_id)
-    im = d.pixel_array
-
-    # Convert from single-channel grayscale to 3-channel RGB
-    im = np.stack([im] * 3, axis=2)
-
-    # Add boxes with random color if present
-    for box in parsed_df[patient_id]['boxes']:
-        rgb = np.floor(np.random.rand(3) * 256).astype('int')
-        im = overlay_box(im=im, box=box, rgb=rgb, stroke=6)
-
-    pylab.imshow(im, cmap=pylab.cm.gist_gray)
-    pylab.axis('off')
-
-def overlay_box(im, box, rgb, stroke=1):
-    """
-    Method to overlay single box on image.
-
-    References:
-    * Exploratory Data Analysis, by Peter Chang, MD.
-        - Available at https://www.kaggle.com/peterchang77/exploratory-data-analysis/notebook
-        - Retrieved 2018-09-19
-    """
-    # Convert coordinates to integers
-    box = [int(b) for b in box]
-
-    # Extract coordinates
-    y1, x1, height, width = box
-    y2 = y1 + height
-    x2 = x1 + width
-
-    im[y1:y1 + stroke, x1:x2] = rgb
-    im[y2:y2 + stroke, x1:x2] = rgb
-    im[y1:y2, x1:x1 + stroke] = rgb
-    im[y1:y2, x2:x2 + stroke] = rgb
-
-    return im
