+++ src/notebooks/model1_maskrcnn/utils.py
@@ -4,14 +4,18 @@
 import os
 import sys
 import random
-import matplotlib.pyplot as plt
 
-sys.path.append(os.path.join('/projects/lungbox'))
-sys.path.append(os.path.join('/projects/lungbox/libs/Mask_RCNN'))
+try:
+    from matplotlib import pyplot as plt
+    _HAS_MATPLOTLIB = True
+except ImportError:
+    _HAS_MATPLOTLIB = False
 
-import libs.Mask_RCNN.mrcnn.model as modellib
-import libs.Mask_RCNN.mrcnn.utils as modelutils
-import libs.Mask_RCNN.mrcnn.visualize as visualize
+script_path = os.path.dirname(os.path.abspath(__file__))
+sys.path.append(script_path + "/Mask_RCNN")
+import Mask_RCNN.mrcnn.model as modellib
+import Mask_RCNN.mrcnn.utils as modelutils
+import Mask_RCNN.mrcnn.visualize as visualize
 
 
 def split_dataset(ids, validation_split=0.2):
@@ -56,6 +60,55 @@ def find_latest_model(model_dir, prefix, inference_config):
     model_path = sorted(fps)[-1]
     print('Found model {}'.format(model_path))
     return model_path
+
+
+def draw(parsed_df, patient_id):
+    """
+    Method to draw single patient with bounding box(es) if present.
+
+    References:
+    * Exploratory Data Analysis, by Peter Chang, MD.
+        - Available at https://www.kaggle.com/peterchang77/exploratory-data-analysis/notebook
+        - Retrieved 2018-09-19
+    """
+    d = get_s3_dcm(patient_id)
+    im = d.pixel_array
+
+    # Convert from single-channel grayscale to 3-channel RGB
+    im = np.stack([im] * 3, axis=2)
+
+    # Add boxes with random color if present
+    for box in parsed_df[patient_id]['boxes']:
+        rgb = np.floor(np.random.rand(3) * 256).astype('int')
+        im = overlay_box(im=im, box=box, rgb=rgb, stroke=6)
+
+    pylab.imshow(im, cmap=pylab.cm.gist_gray)
+    pylab.axis('off')
+
+
+def overlay_box(im, box, rgb, stroke=1):
+    """
+    Method to overlay single box on image.
+
+    References:
+    * Exploratory Data Analysis, by Peter Chang, MD.
+        - Available at https://www.kaggle.com/peterchang77/exploratory-data-analysis/notebook
+        - Retrieved 2018-09-19
+    """
+    # Convert coordinates to integers
+    box = [int(b) for b in box]
+
+    # Extract coordinates
+    y1, x1, height, width = box
+    y2 = y1 + height
+    x2 = x1 + width
+
+    im[y1:y1 + stroke, x1:x2] = rgb
+    im[y2:y2 + stroke, x1:x2] = rgb
+    im[y1:y2, x1:x1 + stroke] = rgb
+    im[y1:y2, x2:x2 + stroke] = rgb
+
+    return im
 
 
 def get_colors_for_class_ids(class_ids):
